# consumer-api/src/main/resources/openapi/consumer-api.yml
openapi: 3.0.3
info:
  title: Readscape-JP Consumer API
  description: 日本語対応書籍販売システムの消費者向けREST API
  version: 1.0.0
  contact:
    name: Readscape-JP API Team
    email: api@readscape.jp
    url: https://readscape.jp/api/docs
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: 開発環境
  - url: https://consumer-api-dev.readscape.jp
    description: テスト環境
  - url: https://consumer-api.readscape.jp
    description: 本番環境
security:
  - BearerAuth: []
  - {}
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT トークンによる認証
  schemas:
...
    ApiErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: エラーコード
        message:
          type: string
          description: エラーメッセージ
        status:
          type: integer
          description: HTTPステータスコード
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          description: リクエストパス
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
      required:
        - error
        - message
        - status
        - timestamp
    FieldError:
      type: object
      properties:
        field:
          type: string
          description: エラーが発生したフィールド名
        message:
          type: string
          description: フィールドエラーメッセージ
    BookSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 書籍ID
        title:
          type: string
          description: 書籍タイトル
        author:
          type: string
          description: 著者名
        price:
          type: integer
          description: 価格（円）
        category:
          type: string
          description: カテゴリー
        rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
          description: 平均評価
        reviewCount:
          type: integer
          description: レビュー数
        imageUrl:
          type: string
          format: uri
          description: 書籍画像URL
        inStock:
          type: boolean
          description: 在庫有無
      required:
        - id
        - title
        - author
        - price
        - category
        - rating
        - reviewCount
        - inStock
...
    PaginationInfo:
      type: object
      properties:
        currentPage:
          type: integer
          description: 現在のページ番号
        totalPages:
          type: integer
          description: 総ページ数
        totalElements:
          type: integer
          format: int64
          description: 総要素数
        size:
          type: integer
          description: ページサイズ
        hasNext:
          type: boolean
          description: 次ページの有無
        hasPrevious:
          type: boolean
          description: 前ページの有無
      required:
        - currentPage
        - totalPages
        - totalElements
        - size
        - hasNext
        - hasPrevious
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ユーザーID
        email:
          type: string
          format: email
          description: メールアドレス
        name:
          type: string
          description: 氏名
        createdAt:
          type: string
          format: date-time
          description: 登録日時
        updatedAt:
          type: string
          format: date-time
          description: 更新日時
      required:
        - id
        - email
        - name
        - createdAt
        - updatedAt
...
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 注文ID
        orderNumber:
          type: string
          description: 注文番号
        status:
          type: string
          enum: [PENDING, CONFIRMED, PROCESSING, SHIPPED, DELIVERED, CANCELLED]
          description: 注文状態
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
          description: 注文アイテム一覧
        totalPrice:
          type: integer
          description: 合計金額（円）
        shippingAddress:
          type: string
          description: 配送先住所
        paymentMethod:
          type: string
          description: 支払い方法
        orderDate:
          type: string
          format: date-time
          description: 注文日時
        estimatedDelivery:
          type: string
          format: date
          description: 配送予定日
      required:
        - id
        - orderNumber
        - status
        - items
        - totalPrice
        - shippingAddress
        - paymentMethod
        - orderDate
    OrderItem:
      type: object
      properties:
        book:
          $ref: '#/components/schemas/BookSummary'
        quantity:
          type: integer
          description: 数量
        price:
          type: integer
          description: 購入時価格（円）
      required:
        - book
        - quantity
        - price
    CreateOrderRequest:
      type: object
      properties:
        paymentMethod:
          type: string
          enum: [credit_card, debit_card, bank_transfer, cash_on_delivery]
          description: 支払い方法
        shippingAddress:
          type: string
          description: 配送先住所
      required:
        - paymentMethod
        - shippingAddress
...
paths:
...
  /users/profile:
    get:
      tags:
        - Users
      summary: プロフィール取得
      description: 認証済みユーザーのプロフィール情報を取得します
      responses:
        '200':
          description: プロフィール取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
  /orders:
    get:
      tags:
        - Orders
      summary: 注文履歴取得
      description: 認証済みユーザーの注文履歴を取得します
      parameters:
        - name: page
          in: query
          description: ページ番号
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: ページサイズ
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: 注文履歴取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  pagination:
                    $ref: '#/components/schemas/PaginationInfo'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

    post:
      tags:
        - Orders
      summary: 注文作成
      description: カートの内容から注文を作成します
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: 注文作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...